Debugged Duplicate & Emulator Issues
I solved the issue of creating a term resulting in what appeared to be two terms, and the same sort of issue for the study sets. Also, I discovered and solved an issue of my emulator not being up to date that caused with testing the program. I also added a page for the Visual study set button to go to that indicates that it is not implemented yet, which users can get out of by pressing the escape button.